+++ .eslintrc.js
@@ -7,7 +7,8 @@ module.exports = {
     "one-var": "off",
     "spaced-comment": ["error", "always", { exceptions: ["-"] }],
     "valid-jsdoc": "off",
-    "prettier/prettier": "warn"
+    "prettier/prettier": "warn",
+    "prefer-const": ["error", { destructuring: "all" }]
   },
   env: {
     es6: true