+++ src/info.js
@@ -173,11 +173,12 @@ export default class Info {
     let intl = false,
       intlTokens = false,
       zones = false,
-      relative = hasRelative();
+      relative = false;
 
     if (hasIntl()) {
       intl = true;
       intlTokens = hasFormatToParts();
+      relative = hasRelative();
 
       try {
         zones =